Added Autocomplete & Cache
The select menu will only allow 25 options, so we needed to go back to a basic slash command. That necessitated autocomplete to act as "options". Since the autocomplete is run for every character typed, and I dont want to hit the server that much, I made a cache for players and teams.